test(auth): split OIDC contract modules (#156)
